pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "throughline"
version = "0.1.0b1"
description = "Persistent long-term memory for Claude Code"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Michael Kupermann", email = "michael@kupermann.com" },
]
keywords = [
    "claude",
    "claude-code",
    "memory",
    "llm",
    "postgresql",
    "pgvector",
    "knowledge-graph",
    "semantic-search",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "psycopg2-binary>=2.9.9",
    "streamlit>=1.32.0",
    "pandas>=2.0.0",
    "plotly>=5.18.0",
    "pyyaml>=6.0",
    "streamlit-agraph>=0.0.45",
    "streamlit-calendar>=1.3.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.30.0",
]
anthropic = [
    "anthropic>=0.40.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "black>=24.0",
]

[project.urls]
Homepage = "https://github.com/mkupermann/throughline"
Repository = "https://github.com/mkupermann/throughline"
Issues = "https://github.com/mkupermann/throughline/issues"
Changelog = "https://github.com/mkupermann/throughline/blob/main/CHANGELOG.md"

[project.scripts]
throughline = "throughline.cli:main"

[tool.setuptools.packages.find]
include = ["throughline*"]
exclude = ["tests*", "scripts*", "gui*", "docs*", "examples*", "skill*", "sql*", "launchd*"]

[tool.setuptools.package-data]
throughline = ["py.typed"]

[tool.ruff]
line-length = 120
target-version = "py310"
extend-exclude = [
    "build",
    "dist",
    ".venv",
    "venv",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # function call as default argument
]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]
include = "\\.pyi?$"
extend-exclude = """
/(
  | build
  | dist
  | \\.venv
  | venv
)/
"""
